Updated and optimized serial threads
cleaned up code for serial thread helper sketches. Now they are
identical except for one line at the top that defines which teensy's
it's going to update.

If the master process is communicating on 1-3, then helper 0 should be
doing 4-6, and helper 1 doing 7-9.

More or less helpers can be configured and used depending on load per
process and how well the cpu handles the delegation.