Schedule system working, reads from CSV
there's still some random non fatal errors popping up but overall the
code is working and the schedule system working too.

A CSV file is downloaded from google drive, fed into a touch designer
network which parses it and converts dates to epoch time.

that txt file is loaded by processing and used as the "playlist". All
timing and scheduling is hardcoded.